requirements.txt
pandas>=2.0.0
requests>=2.31.0
google-cloud-secret-manager>=2.16.0
google-cloud-bigquery>=3.11.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
statsmodels>=0.14.0 